feat: update peer stats only when profile window is open
Polling for the number of peers is now only updated when the user opens
the profile window rather than having a timer constantly updating in the
background.

Got rid of `shouldPoll` because it seems unnecessary with the change in
actually shown.

Added a new event handler to trigger the polling when the window is
how timers are set. The application also doesn't immediately run
`pollStats` when loading the ipfs daemon.