Update conventions link in CONTRIBUTING
In #519, the CONTRIBUTING file suggests adhering to a conventions file when submitting a PR. The link to this file is broken in the Station CONTRIBUTING.md. This PR updates the link. 

As the originally linked page no longer exists, I am unsure if this was the intended link, but based on the now-changed project structure, the new URI is my best guess as to what was intended.