fix: cleanup before starting (fixes #746)
If there is a `config` file, we run `ipfs repo fsck` command to cleanup the lock files. If there is no `config` file, we expect a remote daemon so we do nothing.

`repo fsck` isn't supposed to remove the locks if a daemon is already running so there is no issue (tested on Windows).

I didn't add the code **only** if we got ECONNREFUSED because this way we prevent it already and simplify the code.