chore: deprecate ipfs-on-path (#1768)
This issue relies on brittle scripts which break over time and degrade
user experience. It also introduces maintenance overhead which we don't
have bandwidth to deal with.

This change keeps the feature enabled for existing users, but will hide
menu item for new users, so it is phased out over time.